requirements.txt
# Numerical
numpy
pandas
# tensorflow-gpu

# Helpers
coloredlogs
tqdm
joblib

# Image
Pillow
opencv-python
imutils
dlib
pixelhouse>=0.4.2
sklearn
